requirements.txt
requests==2.22.0
pycryptodome==3.10.1
